SAN-222: Move issuer gateway logic into IG package and create common package for shared services
