+++ .dockerignore
@@ -1,6 +1,7 @@
 *
 
 !assets
+!common
 !config
 !constants
 !dao
