+++ common/services/response_type.go
@@ -1,4 +1,4 @@
-package service
+package services
 
 // ResponseType enumerates the types of authentication supported
 type ResponseType int
