+++ common/services/response_types_test.go
@@ -1,4 +1,4 @@
-package service
+package services
 
 import (
 	"testing"
