+++ service/penalties.go
@@ -1,17 +1,13 @@
 package service
 
 import (
-	"errors"
 	"fmt"
-	"strconv"
 
 	"github.com/companieshouse/chs.go/log"
 	"github.com/companieshouse/penalty-payment-api-core/models"
-	"github.com/companieshouse/penalty-payment-api-core/validators"
+	"github.com/companieshouse/penalty-payment-api/common/services"
 	"github.com/companieshouse/penalty-payment-api/config"
 	"github.com/companieshouse/penalty-payment-api/e5"
-	"github.com/companieshouse/penalty-payment-api/issuer_gateway/api"
-	"github.com/companieshouse/penalty-payment-api/issuer_gateway/types"
 	"github.com/companieshouse/penalty-payment-api/utils"
 )
 
@@ -38,26 +34,21 @@ var getTransactions = func(companyNumber string, companyCode string, penaltyDeta
 	return client.GetTransactions(&e5.GetTransactionsInput{CompanyNumber: companyNumber, CompanyCode: companyCode})
 }
 
-var getAccountPenalties = func(companyNumber string, companyCode string, penaltyDetailsMap *config.PenaltyDetailsMap,
-	allowedTransactionsMap *models.AllowedTransactionMap) (*models.TransactionListResponse, types.ResponseType, error) {
-	return api.AccountPenalties(companyNumber, companyCode, penaltyDetailsMap, allowedTransactionsMap)
-}
-
 // GetPenalties is a function that:
 // 1. makes a request to e5 to get a list of transactions for the specified company
 // 2. takes the results of this request and maps them to a format that the penalty-payment-web can consume
 func GetPenalties(companyNumber string, companyCode string, penaltyDetailsMap *config.PenaltyDetailsMap,
-	allowedTransactionsMap *models.AllowedTransactionMap) (*models.TransactionListResponse, ResponseType, error) {
+	allowedTransactionsMap *models.AllowedTransactionMap) (*models.TransactionListResponse, services.ResponseType, error) {
 	cfg, err := config.Get()
 	if err != nil {
-		return nil, Error, nil
+		return nil, services.Error, nil
 	}
 	client := e5.NewClient(cfg.E5Username, cfg.E5APIURL)
 	e5Response, err := getTransactions(companyNumber, companyCode, penaltyDetailsMap, client)
 
 	if err != nil {
 		log.Error(fmt.Errorf("error getting transaction list: [%v]", err))
-		return nil, Error, err
+		return nil, services.Error, err
 	}
 
 	// Generate the CH preferred format of the results i.e. classify the transactions into payable "penalty" types or
@@ -67,11 +58,11 @@ func GetPenalties(companyNumber string, companyCode string, penaltyDetailsMap *c
 	if err != nil {
 		err = fmt.Errorf("error generating transaction list from the e5 response: [%v]", err)
 		log.Error(err)
-		return nil, Error, err
+		return nil, services.Error, err
 	}
 
 	log.Info("Completed GetPenalties request and mapped to CH penalty transactions", log.Data{"company_number": companyNumber})
-	return generatedTransactionListFromE5Response, Success, nil
+	return generatedTransactionListFromE5Response, services.Success, nil
 }
 
 func generateTransactionListFromE5Response(e5Response *e5.GetTransactionsResponse, companyCode string,
@@ -117,105 +108,4 @@ func generateTransactionListFromE5Response(e5Response *e5.GetTransactionsRespons
 		payableTransactionList.Items = append(payableTransactionList.Items, listItem)
 	}
 	return &payableTransactionList, nil
-}
-
-// MarkTransactionsAsPaid will update the transactions in E5 as paid.
-// resource - is the payable resource from the db representing the penalty(ies)
-// payment - is the information about the payment session
-func MarkTransactionsAsPaid(svc *PayableResourceService, client *e5.Client, resource models.PayableResource, payment validators.PaymentInformation) error {
-	amountPaid, err := strconv.ParseFloat(payment.Amount, 32)
-	if err != nil {
-		log.Error(err, log.Data{"payment_id": payment.Reference, "amount": payment.Amount})
-		return err
-	}
-
-	var transactions []*e5.CreatePaymentTransaction
-
-	for _, t := range resource.Transactions {
-		transactions = append(transactions, &e5.CreatePaymentTransaction{
-			Reference: t.TransactionID,
-			Value:     t.Amount,
-		})
-	}
-
-	// this will be used for the PUON value in E5. it is referred to as paymentId in their spec. X is prefixed to it
-	// so that it doesn't clash with other PUON's from different sources when finance produce their reports - namely
-	// ones that begin with 'LP' which signify penalties that have been paid outside of the digital service.
-	paymentID := "X" + payment.PaymentID
-
-	companyCode, err := getCompanyCodeFromTransaction(resource.Transactions)
-
-	if err != nil {
-		log.Error(err)
-		return err
-	}
-
-	// three http requests are needed to mark a transactions as paid. The process is 1) create the payment, 2) authorise
-	// the payments and finally 3) confirm the payment. if anyone of these fails, the company account will be locked in
-	// E5. Finance have confirmed that it is better to keep these locked as a cleanup process will happen naturally in
-	// the working day.
-	err = client.CreatePayment(&e5.CreatePaymentInput{
-		CompanyCode:   companyCode,
-		CompanyNumber: resource.CompanyNumber,
-		PaymentID:     paymentID,
-		TotalValue:    amountPaid,
-		Transactions:  transactions,
-	})
-
-	if err != nil {
-		if svcErr := svc.RecordE5CommandError(resource, e5.CreateAction); svcErr != nil {
-			log.Error(svcErr, log.Data{"payment_id": payment.PaymentID, "penalty_reference": resource.Reference})
-			return err
-		}
-		logE5Error("failed to create payment in E5", err, resource, payment)
-		return err
-	}
-
-	err = client.AuthorisePayment(&e5.AuthorisePaymentInput{
-		CompanyCode:   companyCode,
-		PaymentID:     paymentID,
-		CardReference: payment.ExternalPaymentID,
-		CardType:      payment.CardType,
-		Email:         payment.CreatedBy,
-	})
-
-	if err != nil {
-		if svcErr := svc.RecordE5CommandError(resource, e5.AuthoriseAction); svcErr != nil {
-			log.Error(svcErr, log.Data{"payment_id": payment.PaymentID, "penalty_reference": resource.Reference})
-			return err
-		}
-		logE5Error("failed to authorise payment in E5", err, resource, payment)
-		return err
-	}
-
-	err = client.ConfirmPayment(&e5.PaymentActionInput{
-		CompanyCode: companyCode,
-		PaymentID:   paymentID,
-	})
-
-	if err != nil {
-		if svcErr := svc.RecordE5CommandError(resource, e5.ConfirmAction); svcErr != nil {
-			log.Error(svcErr, log.Data{"payment_id": payment.PaymentID, "penalty_reference": resource.Reference})
-			return err
-		}
-		logE5Error("failed to confirm payment in E5", err, resource, payment)
-		return err
-	}
-
-	log.Info("marked penalty transaction(s) as paid in E5", log.Data{
-		"penalty_reference": resource.Reference,
-		"payment_id":        payment.PaymentID,
-		"e5_puon":           payment.PaymentID,
-	})
-
-	return nil
-}
-
-func logE5Error(message string, originalError error, resource models.PayableResource, payment validators.PaymentInformation) {
-	log.Error(errors.New(message), log.Data{
-		"penalty_reference": resource.Reference,
-		"payment_id":        payment.PaymentID,
-		"amount":            payment.Amount,
-		"error":             originalError,
-	})
 }
